refactor(SEO): relocate component and update paths
