+++ pages/group/create/index.jsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react';
 import { useRouter } from 'next/router';
 import { useSnackbar } from '@/contexts/Snackbar';
 import useMutation from '@/hooks/useMutation';
-import SEOConfig from '@/shared/components/SEO';
+import SEOConfig from '@/components/SEOConfig';
 import GroupForm from '@/components/Group/Form';
 
 function CreateGroupPage() {
