+++ pages/manage/projects/outcomes/detail.tsx
@@ -3,7 +3,7 @@ import { useRouter } from "next/router";
 import { useMemo, useState } from "react";
 import OutcomeDetail from "@/features/projects/components/OutcomeDetail";
 import { getManageProjectLayout } from "@/layout/features/getProjectLayout";
-import SEOConfig from "@/shared/components/SEO";
+import SEOConfig from "@/components/SEOConfig";
 import { useProject } from "@/services/projects";
 import { useProjectOutcome } from "@/features/projects/hooks/outcome";
 import { parseToNumber, parseToString } from "@/utils/helper";
