+++ pages/manage/projects/outcomes/index.tsx
@@ -1,7 +1,7 @@
 import toast from "react-hot-toast";
 import { useMemo, useState } from "react";
 import { useRouter } from "next/router";
-import SEOConfig from "@/shared/components/SEO";
+import SEOConfig from "@/components/SEOConfig";
 import {
   ContentCard,
   OutcomeCreateModal,
